Compiler warnings

+++ src/core/distributions/phylogenetics/BranchHeterogeneousCharEvoModel.h
@@ -145,7 +145,7 @@ template<class charType, class treeType>
 void RevBayesCore::BranchHeterogeneousCharEvoModel<charType, treeType>::computeInternalNodeLikelihood(const TopologyNode &node, size_t left, size_t right) {    
     
     // get the index in the node sequence
-    unsigned int nodeIndex = node.getIndex();
+    size_t nodeIndex = node.getIndex();
     
     // compute the transition probability matrix
     rateMatrix->getValue()[nodeIndex].calculateTransitionProbabilities( node.getBranchLength(), transitionProbabilities);
@@ -188,7 +188,7 @@ template<class charType, class treeType>
 void RevBayesCore::BranchHeterogeneousCharEvoModel<charType, treeType>::computeTipLikelihood(const TopologyNode &node) {    
     
     // get the index in the node sequence
-    unsigned int nodeIndex = node.getIndex();
+    size_t nodeIndex = node.getIndex();
     
     std::vector<std::vector<double> > &p_node = this->partialLikelihoods[this->activeLikelihood[nodeIndex]][nodeIndex];
